fixed compile errors
