+++ src/genn/genn/code_generator/backendCUDAHIP.cc
@@ -1,4 +1,4 @@
-#include "backend.h"
+#include "code_generator/backendCUDAHIP.h"
 
 // Standard C++ includes
 #include <algorithm>
@@ -15,9 +15,6 @@
 #include "code_generator/modelSpecMerged.h"
 #include "code_generator/standardLibrary.h"
 
-// CUDA backend includes
-#include "utils.h"
-
 using namespace GeNN;
 using namespace GeNN::CodeGenerator;
 
@@ -26,6 +23,11 @@ using namespace GeNN::CodeGenerator;
 //--------------------------------------------------------------------------
 namespace
 {
+const EnvironmentLibrary::Library backendFunctions = {
+    {"clz", {Type::ResolvedType::createFunction(Type::Int32, {Type::Uint32}), "__clz($(0))"}},
+    {"atomic_or", {Type::ResolvedType::createFunction(Type::Void, {Type::Uint32.createPointer(), Type::Uint32}), "atomicOr($(0), $(1))"}},
+};
+
 //--------------------------------------------------------------------------
 // Timer
 //--------------------------------------------------------------------------
@@ -176,32 +178,14 @@ size_t getGroupStartIDSize(const std::vector<T> &mergedGroups)
                                return acc + (sizeof(unsigned int) * ng.getGroups().size());
                            });
 }
-//-----------------------------------------------------------------------
-const EnvironmentLibrary::Library &getRNGFunctions(const Type::ResolvedType &precision)
-{
-    if(precision == Type::Float) {
-        return floatRandomFunctions;
-    }
-    else {
-        assert(precision == Type::Double);
-        return doubleRandomFunctions;
-    }
-}
 }   // Anonymous namespace
 
 
 //--------------------------------------------------------------------------
 // GeNN::CodeGenerator::BackendCUDAHIP
 //--------------------------------------------------------------------------
-namespace GeNN::CodeGenerator::BackendCUDAHIP
+namespace GeNN::CodeGenerator
 {
-bool BackendCUDAHIP::areSharedMemAtomicsSlow() const
-{
-    // If device is older than Maxwell, we shouldn't use shared memory as atomics are emulated
-    // and actually slower than global memory (see https://devblogs.nvidia.com/gpu-pro-tip-fast-histograms-using-shared-atomics-maxwell/)
-    return (getChosenCUDADevice().major < 5);
-}
-//--------------------------------------------------------------------------
 std::string BackendCUDAHIP::getThreadID(unsigned int axis) const
 {
     switch(axis) {
@@ -227,26 +211,6 @@ std::string BackendCUDAHIP::getBlockID(unsigned int axis) const
         return "blockIdx.z"; 
     default:
         assert(false);
-    }
-}
-//--------------------------------------------------------------------------
-std::string BackendCUDAHIP::getAtomic(const Type::ResolvedType &type, AtomicOperation op, AtomicMemSpace) const
-{
-    // If operation is an atomic add
-    if(op == AtomicOperation::ADD) {
-        if(((getChosenCUDADevice().major < 2) && (type == Type::Float))
-           || (((getChosenCUDADevice().major < 6) || (getRuntimeVersion() < 8000)) && (type == Type::Double)))
-        {
-            return "atomicAddSW";
-        }
-
-        return "atomicAdd";
-    }
-    // Otherwise, it's an atomic or
-    else {
-        assert(op == AtomicOperation::OR);
-        assert(type == Type::Uint32 || type == Type::Int32);
-        return "atomicOr";
     }
 }
 //--------------------------------------------------------------------------
@@ -285,11 +249,6 @@ std::string BackendCUDAHIP::genGlobalRNGSkipAhead(CodeStream &os, const std::str
     os << "curandStatePhilox4_32_10_t localRNG = d_rng;" << std::endl;
     os << "skipahead_sequence((unsigned long long)" << sequence << ", &localRNG);" << std::endl;
     return "localRNG";
-}
-//--------------------------------------------------------------------------
-Type::ResolvedType BackendCUDAHIP::getPopulationRNGType() const
-{
-    return CURandState;
 }
 //--------------------------------------------------------------------------
 void BackendCUDAHIP::genNeuronUpdate(CodeStream &os, ModelSpecMerged &modelMerged, BackendBase::MemorySpaces &memorySpaces, 
@@ -785,7 +744,7 @@ void BackendCUDAHIP::genCustomUpdate(CodeStream &os, ModelSpecMerged &modelMerge
             }
 
             // If NCCL reductions are enabled
-            if(getPreferences<Preferences>().enableNCCLReductions) {
+            if(getPreferences<PreferencesCUDAHIP>().enableNCCLReductions) {
                 // Loop through custom update host reduction groups and
                 // generate reductions for those in this custom update group
                 modelMerged.genMergedCustomUpdateHostReductionGroups(
@@ -1052,7 +1011,7 @@ void BackendCUDAHIP::genInit(CodeStream &os, ModelSpecMerged &modelMerged, Backe
     os << initStream.str();
 }
 //--------------------------------------------------------------------------
-void BackendCUDAHIP::genDefinitionsPreamble(CodeStream &os, const ModelSpecMerged &) const
+void BackendCUDAHIP::genDefinitionsPreamble(CodeStream &os, const ModelSpecMerged &modelMerged) const
 {
     os << "// Standard C++ includes" << std::endl;
     os << "#include <random>" << std::endl;
@@ -1063,92 +1022,8 @@ void BackendCUDAHIP::genDefinitionsPreamble(CodeStream &os, const ModelSpecMerge
     os << "#include <cassert>" << std::endl;
     os << "#include <cstdint>" << std::endl;
 
-    os << "// CUDA includes" << std::endl;
-    os << "#include <curand_kernel.h>" << std::endl;
-    if(getRuntimeVersion() >= 9000) {
-        os <<"#include <cuda_fp16.h>" << std::endl;
-    }
+    genDefinitionsPreambleInternal(os, modelMerged);
 
-    // If NCCL is enabled
-    if(getPreferences<PreferencesCUDAHIP>().enableNCCLReductions) {
-        // Include NCCL header
-        os << "#include <nccl.h>" << std::endl;
-        os << std::endl;
-
-        os << std::endl;
-        os << "// ------------------------------------------------------------------------" << std::endl;
-        os << "// Helper macro for error-checking NCCL calls" << std::endl;
-        os << "#define CHECK_NCCL_ERRORS(call) {\\" << std::endl;
-        os << "    ncclResult_t error = call;\\" << std::endl;
-        os << "    if (error != ncclSuccess) {\\" << std::endl;
-        os << "        throw std::runtime_error(__FILE__\": \" + std::to_string(__LINE__) + \": nccl error \" + std::to_string(error) + \": \" + ncclGetErrorString(error));\\" << std::endl;
-        os << "    }\\" << std::endl;
-        os << "}" << std::endl;
-
-        // Define NCCL ID and communicator
-        os << "extern ncclUniqueId ncclID;" << std::endl;
-        os << "extern ncclComm_t ncclCommunicator;" << std::endl;
-
-        // Export ncclGetUniqueId function
-        os << "extern \"C\" {" << std::endl;
-        os << "EXPORT_VAR const size_t ncclUniqueIDSize;" << std::endl;
-        os << "EXPORT_FUNC void ncclGenerateUniqueID();" << std::endl;
-        os << "EXPORT_FUNC void ncclInitCommunicator(int rank, int numRanks);" << std::endl;
-        os << "EXPORT_FUNC unsigned char *ncclGetUniqueID();" << std::endl;
-        os << "}" << std::endl;
-    }
-
-    os << std::endl;
-    os << "// ------------------------------------------------------------------------" << std::endl;
-    os << "// Helper macro for error-checking CUDA calls" << std::endl;
-    os << "#define CHECK_CUDA_ERRORS(call) {\\" << std::endl;
-    os << "    cudaError_t error = call;\\" << std::endl;
-    os << "    if (error != cudaSuccess) {\\" << std::endl;
-    os << "        throw std::runtime_error(__FILE__\": \" + std::to_string(__LINE__) + \": cuda error \" + std::to_string(error) + \": \" + cudaGetErrorString(error));\\" << std::endl;
-    os << "    }\\" << std::endl;
-    os << "}" << std::endl;
-    os << std::endl;
-
-
-    // If device is older than SM 6 or we're using a version of CUDA older than 8
-    if ((getChosenCUDADevice().major < 6) || (getRuntimeVersion() < 8000)){
-        os << "// software version of atomic add for double precision" << std::endl;
-        os << "__device__ inline double atomicAddSW(double* address, double val)";
-        {
-            CodeStream::Scope b(os);
-            os << "unsigned long long int* address_as_ull = (unsigned long long int*)address;" << std::endl;
-            os << "unsigned long long int old = *address_as_ull, assumed;" << std::endl;
-            os << "do";
-            {
-                CodeStream::Scope b(os);
-                os << "assumed = old;" << std::endl;
-                os << "old = atomicCAS(address_as_ull, assumed, __double_as_longlong(val + __longlong_as_double(assumed)));" << std::endl;
-            }
-            os << "while (assumed != old);" << std::endl;
-            os << "return __longlong_as_double(old);" << std::endl;
-        }
-        os << std::endl;
-    }
-
-    // If we're using a CUDA device with SM < 2
-    if (getChosenCUDADevice().major < 2) {
-        os << "// software version of atomic add for single precision float" << std::endl;
-        os << "__device__ inline float atomicAddSW(float* address, float val)" << std::endl;
-        {
-            CodeStream::Scope b(os);
-            os << "int* address_as_ull = (int*)address;" << std::endl;
-            os << "int old = *address_as_ull, assumed;" << std::endl;
-            os << "do";
-            {
-                CodeStream::Scope b(os);
-                os << "assumed = old;" << std::endl;
-                os << "old = atomicCAS(address_as_ull, assumed, __float_as_int(val + __int_as_float(assumed)));" << std::endl;
-            }
-            os << "while (assumed != old);" << std::endl;
-            os << "return __int_as_float(old);" << std::endl;
-        }
-        os << std::endl;
-    }
     os << std::endl;
     os << "template<typename RNG>" << std::endl;
     os << "__device__ inline float exponentialDistFloat(RNG *rng)";
@@ -1408,7 +1283,7 @@ void BackendCUDAHIP::genRunnerPreamble(CodeStream &os, const ModelSpecMerged&) c
 #endif
 
      // If NCCL is enabled
-    if(getPreferences<Preferences>().enableNCCLReductions) {
+    if(getPreferences<PreferencesCUDAHIP>().enableNCCLReductions) {
         // Define NCCL ID and communicator
         os << "ncclUniqueId ncclID;" << std::endl;
         os << "ncclComm_t ncclCommunicator;" << std::endl;
@@ -1446,7 +1321,7 @@ void BackendCUDAHIP::genAllocateMemPreamble(CodeStream&, const ModelSpecMerged&)
 void BackendCUDAHIP::genFreeMemPreamble(CodeStream &os, const ModelSpecMerged&) const
 {
     // Free NCCL communicator
-    if(getPreferences<Preferences>().enableNCCLReductions) {
+    if(getPreferences<PreferencesCUDAHIP>().enableNCCLReductions) {
         os << "CHECK_NCCL_ERRORS(ncclCommDestroy(ncclCommunicator));" << std::endl;
     }
 }
@@ -1588,7 +1463,7 @@ void BackendCUDAHIP::genAssert(CodeStream &os, const std::string &condition) con
     os << "assert(" << condition << ");" << std::endl;
 }
 //--------------------------------------------------------------------------
-Backend::MemorySpaces BackendCUDAHIP::getMergedGroupMemorySpaces(const ModelSpecMerged &modelMerged) const
+BackendCUDAHIP::MemorySpaces BackendCUDAHIP::getMergedGroupMemorySpaces(const ModelSpecMerged &modelMerged) const
 {
     // Get size of update group start ids (constant cache is precious so don't use for init groups
     const size_t groupStartIDSize = (getGroupStartIDSize(modelMerged.getMergedNeuronUpdateGroups()) +
@@ -1602,24 +1477,7 @@ Backend::MemorySpaces BackendCUDAHIP::getMergedGroupMemorySpaces(const ModelSpec
 
     // Return available constant memory and to
     return {{"__device__ __constant__", (groupStartIDSize > getChosenDeviceSafeConstMemBytes()) ? 0 : (getChosenDeviceSafeConstMemBytes() - groupStartIDSize)},
-            {"__device__", m_ChosenDevice.totalGlobalMem}};
-}
-//--------------------------------------------------------------------------
-boost::uuids::detail::sha1::digest_type BackendCUDAHIP::getHashDigest() const
-{
-    boost::uuids::detail::sha1 hash;
-
-    // Update hash was name of backend
-    Utils::updateHash("CUDA", hash);
-    
-    // Update hash with chosen device ID and kernel block sizes
-    Utils::updateHash(m_ChosenDeviceID, hash);
-    Utils::updateHash(getKernelBlockSize(), hash);
-
-    // Update hash with preferences
-    getPreferences<Preferences>().updateHash(hash);
-
-    return hash.get_digest();
+            {"__device__", getDeviceMemoryBytes()}};
 }
 //-----------------------------------------------------------------------
 std::string BackendCUDAHIP::getNCCLReductionType(VarAccessMode mode) const
@@ -1666,22 +1524,4 @@ std::string BackendCUDAHIP::getNCCLType(const Type::ResolvedType &type) const
         throw std::runtime_error("Data type '" + type.getName() + "' unsupported by NCCL");
     }
 }
-//--------------------------------------------------------------------------
-void BackendCUDAHIP::genKernelDimensions(CodeStream &os, Kernel kernel, size_t numThreadsX, size_t batchSize, size_t numBlockThreadsY) const
-{
-    // Calculate grid size
-    const size_t gridSize = ceilDivide(numThreadsX, getKernelBlockSize(kernel));
-    assert(gridSize < (size_t)getChosenCUDADevice().maxGridSize[0]);
-    assert(numBlockThreadsY < (size_t)getChosenCUDADevice().maxThreadsDim[0]);
-
-    os << "const dim3 threads(" << getKernelBlockSize(kernel) << ", " << numBlockThreadsY << ");" << std::endl;
-    if(numBlockThreadsY > 1) {
-        assert(batchSize < (size_t)getChosenCUDADevice().maxThreadsDim[2]);
-        os << "const dim3 grid(" << gridSize << ", 1, " << batchSize << ");" << std::endl;
-    }
-    else {
-        assert(batchSize < (size_t)getChosenCUDADevice().maxThreadsDim[1]);
-        os << "const dim3 grid(" << gridSize << ", " << batchSize << ");" << std::endl;
-    }
-}
-}   // namespace GeNN::CodeGenerator::CUDA
+}   // namespace GeNN::CodeGenerator
